+++ backend/src/main/java/kernel360/techpick/feature/tag/service/TagService.java
@@ -10,11 +10,11 @@ import kernel360.techpick.core.exception.feature.tag.ApiTagException;
 import kernel360.techpick.core.exception.feature.user.ApiUserException;
 import kernel360.techpick.core.model.tag.Tag;
 import kernel360.techpick.feature.pick.repository.PickTagRepository;
+import kernel360.techpick.feature.tag.model.TagMapper;
+import kernel360.techpick.feature.tag.model.TagProvider;
 import kernel360.techpick.feature.tag.model.dto.TagCreateRequest;
 import kernel360.techpick.feature.tag.model.dto.TagResponse;
 import kernel360.techpick.feature.tag.model.dto.TagUpdateRequest;
-import kernel360.techpick.feature.tag.model.mapper.TagMapper;
-import kernel360.techpick.feature.tag.model.provider.TagProvider;
 import kernel360.techpick.feature.tag.validator.TagValidator;
 import lombok.RequiredArgsConstructor;
